+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/RootNodeTest.java
@@ -173,7 +173,7 @@ public class RootNodeTest {
         }
 
         @Override
-        public boolean shouldCaptureFrames() {
+        public boolean isCaptureFramesForTrace() {
             return this.shouldCaptureFrames;
         }